+++ tests/test_session.py
@@ -27,12 +27,10 @@ def test_search_get_db(notion: uno.Session) -> None:
 @pytest.mark.vcr()
 def test_whoami_get_user(notion: uno.Session) -> None:
     me = notion.whoami()
-    assert me.name == 'Github Unittests'
     user = notion.get_user(me.id)
     assert user.id == me.id
     user = notion.get_user(me.id, use_cache=False, raise_on_unknown=False)
-    assert user.name == 'Github Unittests'
-    notion.get_user('645e79dd-3e43-40de-9d51-39357c1c427f', use_cache=False)
+    assert user == me
     with pytest.raises(UnknownUserError):
         unknown_id = '745e79dd-3e43-40de-9d51-39357c1c428f'
         notion.get_user(unknown_id, use_cache=False)